tour_agent/requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
langchain-openai
langchain-core
